Assignment by s...e/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
ansible-core==2.13.13
anyio==4.5.2
apturl==0.5.2
async-timeout==5.0.1
attrs==23.2.0
Automat==22.10.0
bcrypt==3.1.7
beautifulsoup4==4.12.3
blinker==1.4
Brlapi==0.7.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==23.10.4
contourpy==1.1.1
cryptography==41.0.7
cssselect==1.2.0
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.16
defer==1.0.6
distro==1.9.0
distro-info==0.23+ubuntu1.1
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.115.8
fasteners==0.14.1
ffmpeg-python==0.2.0
filelock==3.13.1
fonttools==4.47.2
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2023.12.2
future==0.18.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.14.0
httptools==0.1.2
httpx==0.28.1
hyperlink==21.0.0
idna==2.8
ImageHash==4.3.1
imap-tools==1.6.0
importlib-resources==6.1.1
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
Jinja2==3.1.3
jiter==0.8.2
jmespath==1.0.1
joblib==1.3.2
keyring==18.0.1
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==5.1.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.3
matplotlib==3.7.4
monotonic==1.5
mpmath==1.3.0
multidict==6.1.0
mutagen==1.47.0
netifaces==0.10.4
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
olefile==0.46
openai==1.58.1
opencv-python==4.9.0.80
openpyxl==3.1.2
opera-tosca-parser==0.2.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.0.3
paramiko==2.6.0
parsel==1.8.1
pbr==6.0.0
pexpect==4.6.0
Pillow==7.0.0
propcache==0.2.0
Protego==0.3.0
protobuf==3.6.1
psutil==5.5.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.16.2
pycparser==2.21
pycryptodomex==3.20.0
pycups==1.9.73
pydantic==2.10.4
pydantic-core==2.27.2
PyDispatcher==2.0.7
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.1.1
PyNaCl==1.3.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
PyQt5==5.14.1
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-multipart==0.0.20
pytube==15.0.0
pytz==2023.3.post1
PyWavelets==1.4.1
pyxdg==0.26
PyYAML==5.3.1
queuelib==1.6.2
reportlab==3.5.34
requests==2.32.3
requests-file==1.5.1
requests-unixsocket==0.2.0
resolvelib==0.8.1
scikit-learn==1.3.2
scipy==1.10.1
Scrapy==2.11.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
selenium==4.25.0
service-identity==24.1.0
shtab==1.7.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.38
starlette==0.44.0
sympy==1.12
systemd-python==234
testresources==2.0.1
threadpoolctl==3.2.0
tldextract==5.1.1
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.66.4
trio==0.26.2
trio-websocket==0.11.1
triton==2.1.0
Twisted==22.10.0
typing-extensions==4.12.2
tzdata==2023.4
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==2.2.2
usb-creator==0.3.7
uvicorn==0.11.3
uvloop==0.21.0
w3lib==2.1.2
wadllib==1.3.3
websocket-client==1.8.0
websockets==13.0
wsproto==1.2.0
xkit==0.0.0
xlrd==2.0.1
yarl==1.15.2
yt-dlp==2024.8.6
zipp==3.17.0
zope.interface==6.1
